memalloc: page_size is size_t instead of uint32; fixed test_pages
page_size changed to the size_t everywhere
intrusive list link is a first member in page_info struct
fixed crash in test_pages unit test on test failures

+++ src/cachelot/memalloc.h
@@ -125,7 +125,7 @@ namespace cachelot {
         // total amount of memory
         const size_t arena_size;
         // size of the single page
-        const uint32 page_size;
+        const size_t page_size;
         // pointer to the memory arena
         std::unique_ptr<void, decltype(&std::free)> m_arena;
         // logical pages
